+++ src/partials/Chip.svelte
@@ -1,6 +1,7 @@
 <script lang="ts">
   import cx from "classnames"
 
+  export let pad = false
   export let dark = true
   export let light = false
   export let danger = false
@@ -8,13 +9,14 @@
   export let small = false
   export let onRemove = null
 
-  const className = cx($$props.class, "inline-block rounded-full border border-solid cy-chip", {
+  $: className = cx($$props.class, "inline-block rounded-full border border-solid cy-chip", {
     "border-neutral-800": light,
     "border-neutral-100": dark,
     "!border-danger": danger,
     "!border-warning": warning,
     "py-1 px-2": !small,
     "py-0.5 px-2 text-sm": small,
+    "mb-1 mr-1": pad,
   })
 </script>